requirements.txt
flask
requests
python-consul
pika